Link Payment Form Middlware To Redux
This commit links the payment form middleware to redux so it get runs on every update. The middleware only take effect when the action is EDIT_PROFILE and the field to edit is EDIT_PAYMENT.
As of right now, only the email and cvv regexes have been defined, so the rest of the fields need to be added. Also, the errors object needs to get set properly in the reducer and shown visually.

Various linter errors were also fixed.