Add Profile List Reducer
This commit adds a profile list reducer to encompass actions that affect the profiles array. In order to make this work, a new customProfileReducer
had to be created to only change the customProfile when the actions id given was null. Non-null ids passed to the edit action are now able to edit
an existing profile in the profiles array.

Various syntax changes were added to silence linter warnings.