Visually Change Billing Info when billingMatchingShipping changes
This commit adjust the UI to visually show the shipping information in the billing locationFields Component when the billingMatchesShipping field is
true. This allows a user to visually see the billing info change to match the shipping info in addition to the fields being disabled. This does not affect
actions since the change in values only happens when all the fields are disabled.

This is also only a visual change, so when the billingMatchesShipping field is set to false, the original billing information is shown again and the fields
are activated for editing.