Implement Profile Loading
This commit adds the ability to load profiles from the states list of profiles. In order to temporarily short circuit the missing
implementation of setting the selected profile, this is stored on the components state, rather than the redux state. This will be fixed
later, but was done to make the loading action work.