Implement Update Profile
This commit adds the ability for profiles that already exist (and have been loaded) to update the profiles list, rather than get added as a new profile.
This means that a profile can be loaded, edited, then saved, without adding a new profile. Instead the existing profile in the profiles list will be updated
to match the changes that were made.