Core engine (#7)
* styling prefs updated

* Work on clearing validation errors and attempt to make shipping and billing the same component

* Moved shipping and location into one component that stasifies both

* Cleanup

* All fields clearing validation errors when reinputed and removed internal state management in locationFields and paymentFields

* More work on profiles

* navbar updates

* Profiles add countires and states

* bug fix

* Profiles can now get profile information from backend profiles endpoint

* core engine for KITH.COM started

* stash

* Added loading of profiles, however, currently the page renavigates to profiles after the load

* Bug fix where load would submit the page form

* carting and checkout added

* updated dependencies

* little fucking progress. fml I hate this project someone kill me

* little more progress

* code cleanup and link builder integration into discord bot

* updating readme..

* added link builder

* updated findItem

* animated logo files | NEED FIXED in Navbar

* buggy push, navbar needs fixed

* need to fix bodymovin_light.min.js >> bodymovin.min.js

* updated tasks backend services

* untracked files

* updated logo animation .. buggy still

* updated project structure

* link core -> frontend

* fin

* Navbar updates for animation

* added eslint

* find item optimization

* bug fix with UI

* core link + logo animation (size needs adjusted still..)

* Add Redux to Frontend

* Begin Adding Actions and Reducers

This commit starts to add redux actions and reducers for the
app, starting with the profiles components.

* Implement Location Field Actions

This commit updates the Location Field component to use redux
actions instead of modifying the location prop and calling an
onChange handler.

* Add Location Reducer

This commit adds a location reducer and adds it to a top level reducer

* Restructured Actions and Reducers

* Initial React Redux Linking

* task row added when created

* Link Shipping Entry to Redux State

* Link Billing Entry to Redux State

* Link Payment Entry to Redux State

* temp

* temp

* cleanup

* fixed project setup

* More Redux Restructuring

The top level action and reducer files have been moved to the top of the state folder hierarchy. This reduces a folder level and make more sense when
importing it across the app. All view components have been changed to import the top level actions import instead of sub-levels. Import paths have all
been changed to properly compile and run.

* Link Billing Matches Shipping to Redux State

* Link Profile Name to Redux State

* Adjust Action Values to Match State Names

This commit changes the value of the various action fields in order to match the cases of the state tree. This should
ease some confusion and allows us to better correlate actions to state changes.

* Create Stubbed Middleware for Payment Validation

This commit creates stubs for adding payment validation to the middleware stack. The validation still needs to be done and
a function should be created to create a store and add middleware to the stack.

* server ui updated

* settings/server page updates

* settings/server page updates

* icons MWL included

* Link Payment Form Middlware To Redux

This commit links the payment form middleware to redux so it get runs on every update. The middleware only take effect when the action is EDIT_PROFILE and the field to edit is EDIT_PAYMENT.
As of right now, only the email and cvv regexes have been defined, so the rest of the fields need to be added. Also, the errors object needs to get set properly in the reducer and shown visually.

Various linter errors were also fixed.

* Update paymentReducer to correctly update state errors

This commit fixes a bug where the state was being updated outside of the reducer and in the middleware (which is against the redux design). This has been fixed and
the action.errors object has been correctly mapped to paymentReducer.

* Visually Show Payment Field Errors

This commit links the payment errors state to the visual indicators. This allows users to know when their entries are valid.

* Add Remaining Payment Form Validators

This commit adds the remaining payment form validators for credit card numbers and credit card expiration dates. This completes the payment form validators and the other form middlewares can be modeled after this.

* close bot attempt

* fixed closeBot()

* updating tasks

* icon files

* temp

* task redux impl.

* task redux impl. pt 2

* create task redux

* Add Location Validators to Middleware Stack

This commit creates Location Middlewares for the Shipping and Billing forms,
then adds them to the middleware stack.

* it's a mess

* it's a mess

* it's a mess

* auth window

* Add Profile List Reducer

This commit adds a profile list reducer to encompass actions that affect the profiles array. In order to make this work, a new customProfileReducer
had to be created to only change the customProfile when the actions id given was null. Non-null ids passed to the edit action are now able to edit
an existing profile in the profiles array.

Various syntax changes were added to silence linter warnings.

* Refactor LocationFields

This commit refactors LocationFields to include a redux container, rather than have specific ones defined for shipping and billing. The new container allows the field and profile to be specified.

* Refactor PaymentFields

This commit refactors PaymentFields to be a redux container by default, rather than have a separate file for it. it also changes the props to
require a profile to be passed in to edit.

* Visually Change Billing Info when billingMatchingShipping changes

This commit adjust the UI to visually show the shipping information in the billing locationFields Component when the billingMatchesShipping field is
true. This allows a user to visually see the billing info change to match the shipping info in addition to the fields being disabled. This does not affect
actions since the change in values only happens when all the fields are disabled.

This is also only a visual change, so when the billingMatchesShipping field is set to false, the original billing information is shown again and the fields
are activated for editing.

* Implement Adding a Profile

This commit adds the ability to add a profile to the profiles list. When this happens, the currentProfile is reset to default.

* Implement Profile Loading

This commit adds the ability to load profiles from the states list of profiles. In order to temporarily short circuit the missing
implementation of setting the selected profile, this is stored on the components state, rather than the redux state. This will be fixed
later, but was done to make the loading action work.

* Rename Select to Load (Action)

This commit changes the current instances of the select action to a load action. This is what is actually meant by this action, whereas the real select
action will update the selectedProfile state variable (rather than the current profile). This refactor removes the possible name confusion between
the actions.

* Implement Select Action

This commit integrates the select action of the profile list into the redux state tree. A new reducer was added to respond to the select action.

* Implement Update Profile

This commit adds the ability for profiles that already exist (and have been loaded) to update the profiles list, rather than get added as a new profile.
This means that a profile can be loaded, edited, then saved, without adding a new profile. Instead the existing profile in the profiles list will be updated
to match the changes that were made.

* core changes

* electron js tweak

* added server/settings redux

* broke some shit