Move API calls to middleware
This commit moves the API calls necessary for manipulating profiles to a
separate middleware. At this point, the calls are not made, but a mocked
call returning the expected response after a 1 second delay is made.

More testing of the api call will be done to first before we can uncomment
the api call.