Add multi-window support
This commit allows multiple windows to launch by using a workaround
to bypass webpacks bundling. A comment enables require to be loaded
on the window and then allows electron to be loaded.

Using the ipc functions from electron, we can send messages to the
main electron process, which can then use the window manager to
perform all adding/closing of windows.

NOTE: This breaks compatibility with the web-app being loaded in
the browser, since the browser has no notion of the electron
package. The electron app will run fine however.