Fix Server Filter
This commit fixes the fields for filter. The server size is disabled until the server type is set. Once the server type is set, the server size
options are filtered based on the type. Changing the server type will also reset the server size (if it is a different type). This forces the user
the reselect the server size before continuing.