Refactor Server State
This commit starts a refactor of the Server state. This name is a
misnomer since it doesn't represent a Server, but rather info about one
main server and optional proxy server. So, this portion of the state
tree has been renamed to "serverInfo"

Further, this commit removes actions that are unnecessary.