Add Create/Destroy Server Thunks
This commit adds create and destroy actions implemented using thunks.
For now, the bindings are created and a temporary promise is used to
verify the action is called appropriately (using redux devtools).

This commit also adds an error action to the serverActions. This is used
whenever an error occurs for a given action. The current implementation
is to just print and error message to the console, but this will
eventually be changed to notify the user of the error.

The parameters of the create and destroy actions will likely change as
I learn more about what type of information is required to complete the
request of creating a server (I'm just guessing at this point!).