Add Generate/Destroy Proxy Thunks
This commit adds thunks to handle generating and destroying proxies.
This method is mocked out for now, and will be replaced with an actual
request. For now, the wiring of the thunk has been created and the
server reducer correctly responds to the successful action.

The parameters for these functions will likely change as I learn what
specific information is required to successfully perform the request.