Add Validate AWS Thunk
This commit adds a thunk for validating AWS keys. For now, this has been
mocked out to make sure the wiring of the thunk is correct. This will be
replaced with a more permanent request in the future.

Further, this commit adds disabling of the proxy and server control if
AWS credentials have not been validated. Once they are validated, the
other functions are available for use.

The parameters for validation might also change in the future as I
learn what information is required to make this call.