Add Logout AWS Thunk
This commit adds the ability to log out of AWS and reset credentials.
This thunk is just a compound thunk that destroys proxies and servers,
then resets credentials if successful.

This commit also adds a new confirm dialog method to the bridge. This
allows the renderer to create dialogs. This is used when logging out
to confirm that user wants to log out before continuing.