Fix Rerender Bug
This commit fixes a bug that caused the proxy list to lose it's proxy information. The
proxy information still existed on the redux state tree, but when moving to another 
route and back, the ProxyList component was destroyed and reconstructed. This 
caused the internal state to be set to an empty list, rather than the initials proxies
given by the props. 

This commit changes the constructor to set the state of the proxies to the 
input props (or an empty array if no proxies are given). Further, a wrong 
editing state was fixed. (true -> false).