Start Task CSS Reorganization
This commit starts reorganizing the tasks css to fix the 
layout of the tasks table. The first part was moving to 
a consistent naming scheme (BEM) for css classes. 
The next part was renaming class names to match 
the new css. Finally, more ids have been moved to 
using css classnames instead of specific ids to 
promote reusability.