Add npm run sass shortcut
This commit adds an npm shortcut to compile sass into css. Sass was added as a
dev dependency, so it is now invoked through the node_modules folder, rather 
then relying on it being installed globally. Finally, the compile script's file 
permissions are changed to be readable and executable by all users, allowing npm
to run the script.