Start Layout and CSS Rework
this commit starts work on html layout changes they don't rely on the top/left anchors. Instead, a simple grid system using flexboxes was added and used to 
build the layout. 

To aid in comparing to the original, the tasks component was copied. The original is unmodified and instead, a new route/icon was added to render the changed component. This makes it much easier to switch between the original and changed components for comparison. 

Sub-components haven't been changed yet, but will be as more progress is made.