Enable custom paste behavior
This commit changes the paste behavior by intercepting a paste event and capturing only the text (after sanitizing it). This prevents any css or html from making it to the input. 

This works, however, the proxy validator now fails for all pasted lines (except the last) even if they are valid. Once a manual edit (insertion/deletion) is made, all valid pasted lines are correctly counted.