Switch back to regex test instead of match
This commit switches back to regex test to solve the problem of false negatives when pasting in text. This problem is confirmed to be solved on mac, but windows has not been tested yet. 

Comments are added for the paste handler, console log statements are removed. 

Finally, the settings reducer was changed to set the errors object when a specific settings action was received, rather than for all actions. This prevents the clearing of the errors map when another action takes place (or when reloading).