Rework Progress
This commit marks significant progress with the rework. The basic layout is the same now, there are now only minor changes that need to be made to the spacing. 

These fixes will be addressed next, and then the react sub-components can have their css reworked to fit within the reworked layout.