Rework Create Task Layout and Styling
This commit updates the layout and styling of Create Task to fit within the reworked tasks page layout. Styling was updated to use BEM style naming for classes. The create task sub component now renders and functions properly. The layout is very close to the original, save a few spacing differences. These can be addressed further. 

The next task is to rework the log box with a reworked layout.