Implement Upstream Changes
This commit implements the changes from master into the updated layout. This includes adding Select components for the edit menu. Additionally, a sass mixin was added to easily set the same font in multiple places.

Additionally, the placeholder for the Billing Profile in the updated CreateTask now correctly displays when no value is selected.