Use Navbar Redux State
This commit makes use of the newly created Navbar Actions to control navigation. Instead of pushing directly to history, the redux navbar actions are called. 

This commit also makes use of the ROUTES object defined in navbarActions. This will allow consistency between parts of the frontend when specifying a specific route.