Adjust Task Reducer Implementation To Support Task Row Edits
This commit changes the task reducer to support task edits when they come from the edit menu of the task row. Since only certain fields are allowed to be edited when a task has already been created, a separate case was made to handle this in the task reducer. 

This commit also refactors the task list reducer to move the selection of an ID out of the reducer body into its own function. This should support easier adjustments to this method without having to adjust the implementation of reducer specific code. Further, edits to an existing task should happen within an edits object rather than on the full tasks object. This is reflected in the new implementation. 

Finally, this commit adjusts the task list reducer to add an edits object to the task state object. This edits object gets filled when a new task is added to the task list. The edit menu in TaskRow will use this as a source for it's values instead of the tasks actual values.