Refactor Task Row to support Edits
This commit adjusts the implementation of Task Row to link the task edits object to the edit menu.

This commit also heavily refactors the task page by adding sub render methods for individual, repeated objects. Specifically this is for the image buttons that control the task status and the edit menu visibility. 

This commit adds on change handlers for the edit menu fields. Since the reducers are already implemented, the edit menu now saves edits correctly.