Control Username and Password Inputs
This commit fixes an error due to not always controlling the username and password inputs. The fallback empty string is passed in case the username and password are both null. 

This commit also replaces `this.props` references by using object destructuring at the top of functions. This removes a lot of characters from each line and improves readability of the code.