Connect new edit fields to Redux
This commit adds the new edit fields to the redux state and connects the proper values and event handlers to the components. The edit menu functions as expected now. This involves updates to the task actions and task list reducer to properly copy over edit values.  Proper handling of the task edit action was implemented when editing an existing task in the task reducer. Finally, the styling was changed to allow a visual disabled state.