Update Tests to use Mock Redux Store
This commit updates the tests to use the redux-mock-store package to help mock the various components of the redux store. This allows greater flexibility to test redux related structures rather than building custom mocks manually.

This commit also adds the history package as a dev dependency to create in memory history for testing purposes.