Add Profile Definitions Tests
This commit adds unit tests for all the profile definition files. The initial states have been moved to the definitions files instead of the reducers. This seems more natural because now the definitions and initial states are defined in the same place.

A definitions helper util was added to make testing definitions files easier, this will be used in the future for the rest of the definitions.

Finally, the timeouts are removed from the promises of profileActions since they are all zero and are not necessary.