Update Definitions Utils
This commit updates the definitions utils with a new method to test a specific value rather than a key within an object or an array. 

This commit also solves a bug caused by prop types only printing error messages once per error. To get around this, the utils method now has a private index that gets run and appended to the information of each prop types check. This forces prop types to print the warning message for each test regardless of key or value.

Finally this commit removes an unused import from taskList test.