Session Storage progress
This commit adds more progress towards the session storage. Usage is now blocked if the key has already been authenticated. 

A temporary bypass has been added to this to diagnose issues with session storage. The main window appears on successful authentication, but we are unable to load the window contents.