Add Deauth Endpoint
This commit adds a new endpoint for deauthing a user from nebula. The user must pass a valid token to access this end point. If the token is given, the entry in the users table is deleted and the key can be re-registered to another device.

To use this end point, the user must send a DELETE request to `/auth`