Implement Nebula API in frontend
The nebula API authentication flow is now used in the frontend. API calls are made for initial activation, at which point no calls are made until the access token expires.  Deauthing is also implemented, causing a force back to the auth page.

A significant overhaul to the window management in electron.js was made to DRY the window management out. Window management can still be improved in the future.