Change ServerRow definition to Server
This commit changes the name of serverRow to server for the definitions. The properties of the definition aren't really related to a  "row" other than the fact the the ServerRow component uses them. "Server" is therefore a better name for the definition.

This commit also adds initial states for more server definitions.