Add Ability to Register and Get Nodes by Test Id
This commit adds the ability to register a testid onto a node and get it when testing. Test ids are added as an attribute to a node using the key: "data-testid". The method used here requires us to add that attribute onto a node, then call the given "addTestId" function. The environment is checked before adding the test id. This allows test ids to appear when testing but not in development or production.

A simple function was called to search an enzyme wrapper for the resulting test id attribute.