Add App Tests
This commit adds unit tests for the top level App component. The component was updated to pass the given tests.

This commit also removes several files from coverage tracking. This includes old files that are no longer being used in the app, as well as the top level index.js that can't be tested effectively.