Fix Horizontal Scrolling for Create Task Size Select
This commit changes the styling for multi value selects to allow for horizontal scrolling when selected items overflow the select containers area. This is done by setting `flex-wrap` to `nowrap` and `overflow-x` to `scroll` on the value container (when we are a multi value container), and setting the `flex-shrink` to 0 on multi value items.