Update Task Reducer to use Default Values as fallback
This commit updates the task reducer to use default values as the fallback when an empty/falsy value is given as the updated value. This allows fields to get reset back to no value when they get cleared and prevents inputs from not being able to be cleared.

Tests are updated and added to accommodate/preserve this change.