Add Aborted Event Handler
This commit updates the TaskRunner object to accept an instance of the TaskManager that will be controlling it. The constructor now registers for an abort event, so when the event is triggered by the task manager, the proper state is handled.

The internal data of the TaskRunner is now combined into a single context wrapper object. This object is passed to the sub processes so they can receive the information they need to run.