Implement High Level Flow for Monitor
This commit adds the high level flow for how monitoring should work. This includes:
- A race between the three types of product files for product info
- Checking of the status for bans in the responses
- Checking of the product response to see if it is ready for checkout
- Delaying by the monitor or refresh delay if necessary 

The lower level details such as performing/parsing the responses and verifying the product is ready for checkout have not been implemented yet; they are stubbed currently for a WIP look at how the finished implementation will work. The utility functions will get created next to accomplish these tasks, then the stubs will be replaced with the actual implementation.