Add Json Parsing
Parsing in general should take a given task (and proxy), perform a network request for the correct file, then parse the response to make sure the task's product exists. 

This commit adds a parser that works for the `products.json` file of a site. 

More Parsers of different formats will be added.