Add Full Product Info Retrieval
This commit adds a method to the JsonParser, allowing it to retrieve product info for a single product given its url. This method is then used by the Xml and Atom parsers to fill in missing gaps about the matched product, specifically variant information. This method helps standardize the return value from all parsers for further size parsing can take place.

The standalone json and oembed product files are added as assets for reference.