Add Better Utils Parsing Sizes From Variants
This commit adds better utils for getting the size of a variant. The size is available either as a separate option within the variant or within the title. Further, different sites use different option names when storing the site. In order to fix this a new set 
of maps is are available to get the right fields from a variant given the url the variant is from.