Add A Basic Shipping Test Case
This commit adds a temporary basic shipping test Proof of Concept. It demonstrates that getting the shipping form data is possible and values can be injected using query parameters.

To run the script the user will have to update the information specific to a checkout. 

This file is a proof of concept only and will be removed once the changes required are integrated (cookie management).