Move Task Runner Constants to A Utility File
This commit moves the constants associated with the Task Runner to a separate constants file. This was to remove a circular dependency created by the sub processes including the TaskRunner for the constants. Now any class needing the TaskRunner constants can include them separately without worrying about a possible circular dependency.