Add Handler for Captcha Events
This commit adds a handler that gets called when task captcha events are emitted from the TaskManager. These events are related to starting/stopping harvesting tokens for a specific task. These events will call a method on the window manager to handle showing/spawning captcha windows.

The specific event handler body/parameters are subject to change as the task manager development is still in progress.