Connect Captcha Window to Start/Stop Harvest Events
This commit adds more progress to connect the start/stop harvest events to the captcha window. The harvest events flow from the window manager to the captcha window managers to the captcha windows themselves. Handlers were added to listen for the correct events and properly load the captcha when it is necessary.