Add Electron Recommend Security Settings (#108)
* Implement Security Fixes

This commit implements all recommended security fixes from electron with the exception of enabling context isolation (our current frontend heavily relies on this being disabled).

* Fix Captcha Server Errors

This commit updates the captcha server to be an https self-signed server. A certificate handler was added to bypass certificate errors for the captcha server only, other certificate errors will get rejected. The js/html for the captcha page was also updated to adhere to security standards.