Add Dynamic Level Support
This commit updates the create logger to set the starting levels based on the environment. The logger file was moved to the common directory since that is a more suitable location for it.

This commit adds support for dynamically adjusting the level of the created loggers. Levels can be set individually per transport (or all transports together), as well as for a single logger (based on the name) or all loggers.