Add Logger to Task-Runner (#117)
* Add Winston to task-runner

This commit adds winston to the task-runner project and creates a simple function to create a logger.

Initial the logger writes out to the console (when the level if info or higher) writes to a combined log file, and writes to a given file name.

* Use Logger in TaskManager

This commit updates the TaskManager to use a logger instead of using console.log. All previous statements have been updated to use the logger with an appropriate level attached.

* Event message -> payload

This commit refactors the task runner slightly to rename the "message" parameter to a better name: "payload". This is an object that gets passed via events which contains a message as well as other info (errors, etc.). Thus, "payload" better informs developers the purpose of this parameter.

* Add Logger to Task Runner

This commit adds a logger to each task runner and converts console.log statements to log statements. The logger function was also updated with a more consistent format and adjust the default format level to include all levels.

A stray log statement is also removed from TaskManager.

* Ignore Logs folder

This commit updates the frontend gitignore to ignore the logs directory. This will prevent logs in a specific run from getting tracked.

* Add Logger to TaskRunner Classes

This commit uses the context from the TaskRunner to get the TaskRunner's logger and convert console.log statements to log statements.

Some small refactoring was also done to simplify the shipping promise structure.

* Add Logger to Parsers

This commit refactors Parsers to support an optional logger. All console log statements are converted to use the logger if given, otherwise do nothing.

* Add Logger to Utils

This commit adds support of an optional logger to all util methods. If no logger is passed, then no log operations are done. All console.log statements are converted to use the logger.

* Add Dynamic Level Support

This commit updates the create logger to set the starting levels based on the environment. The logger file was moved to the common directory since that is a more suitable location for it.

This commit adds support for dynamically adjusting the level of the created loggers. Levels can be set individually per transport (or all transports together), as well as for a single logger (based on the name) or all loggers.

* Adjust Log Function Called

This commit updates the log function called to use the level specific function instead of the generic log. For places where that's not possible (util methods), the generic log function is used.

* Add Debug Command to Set Log Levels Dynamically

This commit adds a debug command (accessible through ipc) to dynamically set the log levels.

This commit also prevents a bug where the levels for a specific logger would be applied to all future loggers.

* Forward Parser Name Properly

This commit fixes a runtime error due to the super constructor not being called immediately for the parsers. JsonParser now supports a name property upon construction, which the Xml and Atom parsers take advantage of. 

This commit also fixes some smalls syntax errors due to mismatched return types and missing parameters.