Adjust the App Redirect based on the Redux State
This commit updates the App component to redirect to the state's last route rather than always redirect to the tasks page. If no location is set, or the window's location matches the states location. The redirect is defaulted to the tasks page.

This commit also updates the naming of parameters in the navbar actions to be clearer.